First version of the client handler
The client handler is a moderator in the communication between
the client and the sever. It is between the socket connection and
the rest of the server. It consider actions upon incoming massages
from the socket. And responds to the client with information about
the actions and events in the seever.

The client handler is the process which can modify a file though the
corresponding file_controller. The Client handler can also subscribe
on a file so it will get notified of all events and changes that
have been made by it self or an other client.
This is NOT yet implemented.