Fixed #8, It do not revealing the file paths for client any more.
The solution is to always use IDs on both files and directorys.
the client can then ask for content in a directory, which will then
receive a list of content, files or directorys with IDs.
This way dose the client only have a virtual knowledge of the
directory structure, as the server still knows the real structure.